Backend/requirements.txt

# FastAPI backend
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary

# Pydantic for schema validation
pydantic

# Translation service
textblob

# Other utilities
python-dotenv

# Pydantic settings
pydantic-settings

# FastAPI form data support
python-multipart

# Twilio SMS/voice integration
twilio

# Authentication & Security
python-jose
passlib

bcrypt

# Database Migrations
alembic


